+++ Deeploy/Targets/Generic/Parsers.py
@@ -221,6 +221,48 @@ class MaxPoolParser(NodeParser):
         return ctxt, True
 
 
+class MaxPool1DParser(MaxPoolParser):
+
+    def __init__(self):
+        super().__init__()
+
+    def parseNode(self, node: gs.Node) -> bool:
+        ret = super().parseNode(node)
+        wellFormed = False
+        if ret:
+            pads = self.operatorRepresentation['pads']
+            kernel_shape = self.operatorRepresentation['kernel_shape']
+            strides = self.operatorRepresentation['strides']
+            # 1D: pads should be length 2, kernel_shape length 1, strides length 1
+            if len(pads) == 2 and len(kernel_shape) == 1 and len(strides) == 1:
+                wellFormed = True
+                self.operatorRepresentation['padding_y'] = int(pads[0])
+                self.operatorRepresentation['padding_y_right'] = int(pads[1])
+                self.operatorRepresentation['stride_y'] = int(strides[0])
+                self.operatorRepresentation['dim_kernel_y'] = int(kernel_shape[0])
+        return wellFormed
+
+    def parseNodeCtxt(self, ctxt, node, channels_first = True):
+        newCtxt, ret = super().parseNodeCtxt(ctxt, node, channels_first)
+        if ret:
+            data_in = newCtxt.lookup(self.operatorRepresentation['data_in'])
+            data_out = newCtxt.lookup(self.operatorRepresentation['data_out'])
+            self.operatorRepresentation['batch'] = data_in.shape[0]
+            if channels_first:
+                self.operatorRepresentation['ch_im_in'] = data_in.shape[1]
+                self.operatorRepresentation['dim_im_in_y'] = data_in.shape[2]
+                self.operatorRepresentation['ch_im_out'] = data_out.shape[1]
+                self.operatorRepresentation['dim_im_out_y'] = data_out.shape[2]
+            else:
+                self.operatorRepresentation['ch_im_in'] = data_in.shape[2]
+                self.operatorRepresentation['dim_im_in_y'] = data_in.shape[1]
+                self.operatorRepresentation['ch_im_out'] = data_out.shape[2]
+                self.operatorRepresentation['dim_im_out_y'] = data_out.shape[1]
+            if len(data_in.shape) == 3 and len(data_out.shape) == 3:
+                return newCtxt, True
+        return ctxt, False
+
+
 class MaxPool2DParser(MaxPoolParser):
 
     def __init__(self):
@@ -298,7 +340,12 @@ class PadParser(NodeParser):
 
         if ret:
             self.operatorRepresentation['mode'] = node.attrs['mode']
-            self.operatorRepresentation['pads'] = node.attrs['pads']
+
+            try:
+                self.operatorRepresentation['pads'] = [int(p) for p in node.attrs['pads']]
+            except Exception as e:
+                self.operatorRepresentation['pads'] = node.attrs['pads']
+
             self.operatorRepresentation['value'] = node.attrs['value']
 
         return ret
@@ -1325,6 +1372,8 @@ class Conv1DParser(ConvParser):
 
             self.operatorRepresentation['batch'] = data_in.shape[0]
             self.operatorRepresentation['dim_im_in_x'] = 1
+
+            # Necessary, since we use the same Convlayer for all convolutions
             self.operatorRepresentation['dim_im_out_x'] = 1
 
             if channels_first:
@@ -1337,6 +1386,11 @@ class Conv1DParser(ConvParser):
                 self.operatorRepresentation['dim_im_in_y'] = data_in.shape[1]
                 self.operatorRepresentation['ch_im_out'] = data_out.shape[2]
                 self.operatorRepresentation['dim_im_out_y'] = data_out.shape[1]
+
+            self.operatorRepresentation[
+                'batchOffsetIn'] = self.operatorRepresentation['ch_im_in'] * self.operatorRepresentation['dim_im_in_y']
+            self.operatorRepresentation['batchOffsetOut'] = self.operatorRepresentation[
+                'ch_im_out'] * self.operatorRepresentation['dim_im_out_y']
 
             if len(data_in.shape) == 3 and len(weight.shape) == 3:
                 return newCtxt, True
@@ -2136,7 +2190,20 @@ class GenericConv1DParser(Conv1DParser):
 
         if ret:
             inputs = ['data_in', 'weight']
+
+            # Handle bias, if present
+            if len(node.inputs) > 2:
+                inputs.append("bias")
+                self.operatorRepresentation["has_bias"] = 1
+            else:
+                self.operatorRepresentation["has_bias"] = 0
+                self.operatorRepresentation["bias"] = "NULL"
+
             for idx, inputNode in enumerate(node.inputs):
+                if idx >= len(inputs):
+                    raise IndexError(
+                        f"Index {idx} out of range for inputs of length {len(inputs)} in node {inputNode.name}")
+
                 self.operatorRepresentation[inputs[idx]] = ctxt.lookup(inputNode.name).name
 
             return newCtxt, True
@@ -2555,3 +2622,171 @@ class SGDParser(NodeParser):
         self.operatorRepresentation['lr'] = node.attrs['lr']
 
         return ctxt, True
+
+
+class BatchNormParser(NodeParser):
+
+    def __init__(self):
+        super().__init__()
+
+    def parseNode(self, node: gs.Node) -> bool:
+        # Verify the attributes (epsilon is mandatory, momentum and training_mode  are optional)
+        if 'epsilon' not in node.attrs:
+            return False
+        # Common Inputs: 5 (X, scale, B, mean, var)
+        if len(node.inputs) < 5:
+            return False
+
+        # Save the attributes, default values are provided if not present
+        self.operatorRepresentation['epsilon'] = node.attrs.get('epsilon', 1e-5)
+        self.operatorRepresentation['momentum'] = node.attrs.get('momentum', 0.9)
+        self.operatorRepresentation['training_mode'] = node.attrs.get('training_mode', 0)
+
+        return True
+
+    def parseNodeCtxt(self, ctxt, node: gs.Node, channels_first: bool = True):
+        inputs = ['data_in', 'scale', 'bias', 'mean', 'variance']
+        outputs = ['data_out']
+
+        for idx, inputNode in enumerate(node.inputs[:5]):
+            self.operatorRepresentation[inputs[idx]] = ctxt.lookup(inputNode.name).name
+
+        # Output (Y)
+        self.operatorRepresentation[outputs[0]] = ctxt.lookup(node.outputs[0].name).name
+
+        input_shape = ctxt.lookup(node.inputs[0].name).shape
+        # Save input shape information
+        self.operatorRepresentation['batch_size'] = input_shape[0]
+        self.operatorRepresentation['channel_size'] = input_shape[1]
+        self.operatorRepresentation['window_size'] = input_shape[2]
+
+        return ctxt, True
+
+
+class ConvTransposeParser(NodeParser):
+
+    def __init__(self):
+        super().__init__()
+
+    def parseNode(self, node: gs.Node) -> bool:
+        # Extract ONNX attributes with defaults
+        strides = node.attrs.get('strides', [1])
+
+        pads = node.attrs.get('pads', [0, 0])
+        kernel_shape = node.attrs.get('kernel_shape', None)
+        dilations = node.attrs.get('dilations', [1])
+        group = node.attrs.get('group', 1)
+
+        # Check for required attributes
+        wellFormed = (kernel_shape is not None and len(node.outputs) == 1)
+        if wellFormed:
+            self.operatorRepresentation['strides'] = strides
+            self.operatorRepresentation['pads'] = pads
+            self.operatorRepresentation['kernel_shape'] = kernel_shape
+            self.operatorRepresentation['dilations'] = dilations
+            self.operatorRepresentation['group'] = group
+            self.operatorRepresentation['nodeName'] = node.name
+            self.operatorRepresentation['nodeOp'] = node.op
+        return wellFormed
+
+    def parseNodeCtxt(self, ctxt: NetworkContext, node: gs.Node, channels_first: bool = True):
+        # Register buffer names for codegen
+        self.operatorRepresentation['data_in'] = node.inputs[0].name
+        self.operatorRepresentation['weight'] = node.inputs[1].name
+        self.operatorRepresentation['data_out'] = node.outputs[0].name
+        if len(node.inputs) == 3:
+            self.operatorRepresentation['bias'] = node.inputs[2].name
+            self.operatorRepresentation['has_bias'] = "true"
+        else:
+            self.operatorRepresentation['has_bias'] = "false"
+        # Get output shape from context
+        data_out = ctxt.lookup(node.outputs[0].name)
+        out_shape = data_out.shape
+        if len(out_shape) == 3:
+            self.operatorRepresentation['dim_im_out_x'] = out_shape[2]
+        elif len(out_shape) == 4:
+            self.operatorRepresentation['dim_im_out_x'] = out_shape[2]
+            self.operatorRepresentation['dim_im_out_y'] = out_shape[3]
+
+        stride_x, stride_y = 1, 1
+        if "strides" in node.attrs:
+            stride_y = node.attrs["strides"][0]
+            stride_x = node.attrs["strides"][1] if len(node.attrs["strides"]) > 1 else stride_y
+        self.operatorRepresentation["stride_y"] = stride_y
+        self.operatorRepresentation["stride_x"] = stride_x
+
+        if "kernel_shape" in node.attrs:
+            kernel_shape = node.attrs["kernel_shape"]
+            kernel_shape_x = kernel_shape[0]
+            # For 2D, kernel_shape may have two elements
+            kernel_shape_y = kernel_shape[1] if len(kernel_shape) > 1 else kernel_shape_x
+        else:
+            kernel_shape_x = 1
+            kernel_shape_y = 1
+
+        data_in = ctxt.lookup(node.inputs[0].name)
+        data_out = ctxt.lookup(node.outputs[0].name)
+        in_shape = data_in.shape
+        out_shape = data_out.shape
+
+        self.operatorRepresentation['ch_im_in'] = in_shape[1]
+        self.operatorRepresentation['dim_im_in_y'] = in_shape[2]
+        self.operatorRepresentation['ch_im_out'] = out_shape[1]
+        self.operatorRepresentation['dim_im_out_y'] = out_shape[2]
+
+        self.operatorRepresentation[
+            'batchOffsetIn'] = self.operatorRepresentation['ch_im_in'] * self.operatorRepresentation['dim_im_in_y']
+        self.operatorRepresentation[
+            'batchOffsetOut'] = self.operatorRepresentation['ch_im_out'] * self.operatorRepresentation['dim_im_out_y']
+        return ctxt, True
+
+
+class ConvTranspose1DParser(ConvTransposeParser):
+
+    def __init__(self):
+        super().__init__()
+
+    def parseNode(self, node: gs.Node) -> bool:
+        # 1D ConvTranspose expects 3D input/output and 3D weight
+        wellFormed = super().parseNode(node)
+        ret = False
+        if wellFormed:
+            ret = all([
+                # Make sure strides are 2D
+                len(node.attrs['strides']) == 1,
+                len(node.attrs['pads']) == 2,
+                len(node.attrs['dilations']) == 1,
+            ])
+        if ret:
+
+            self.operatorRepresentation['kernel_shape'] = node.attrs['kernel_shape']
+            self.operatorRepresentation['dim_kernel_y'] = int(self.operatorRepresentation['kernel_shape'][0])
+            self.operatorRepresentation['dilation_y'] = int(self.operatorRepresentation['dilations'][0])
+            self.operatorRepresentation['padding_y'] = int(self.operatorRepresentation['pads'][0])
+            self.operatorRepresentation['stride_y'] = int(self.operatorRepresentation['strides'][0])
+
+        return ret
+
+    def parseNodeCtxt(self,
+                      ctxt: NetworkContext,
+                      node: gs.Node,
+                      channels_first: bool = True) -> Tuple[NetworkContext, bool]:
+
+        newCtxt, ret = super().parseNodeCtxt(ctxt, node, channels_first)
+
+        if ret:
+            data_in = newCtxt.lookup(node.inputs[0].name)
+            data_out = newCtxt.lookup(node.outputs[0].name)
+            in_shape = data_in.shape
+            out_shape = data_out.shape
+            self.operatorRepresentation['batch'] = in_shape[0]
+            self.operatorRepresentation['ch_im_in'] = in_shape[1]
+            self.operatorRepresentation['dim_im_in_y'] = in_shape[2]
+            self.operatorRepresentation['ch_im_out'] = out_shape[1]
+            self.operatorRepresentation['dim_im_out_y'] = out_shape[2]
+            self.operatorRepresentation[
+                "batchOffsetIn"] = self.operatorRepresentation["ch_im_in"] * self.operatorRepresentation["dim_im_in_y"]
+            self.operatorRepresentation["batchOffsetOut"] = self.operatorRepresentation[
+                "ch_im_out"] * self.operatorRepresentation["dim_im_out_y"]
+            return newCtxt, True
+        return ctxt, False