+++ Deeploy/Targets/Generic/Templates/FloatConvTemplate.py
@@ -29,3 +29,29 @@ BEGIN_SINGLE_CORE
     }
 END_SINGLE_CORE
 """)
+
+reference1DTemplate = NodeTemplate("""
+<%
+batchOffsetIn = ch_im_in * dim_im_in_y
+batchOffsetOut = ch_im_out * dim_im_out_y
+%>
+    // 1D FP Conv (Name: ${nodeName}, Op: ${nodeOp})
+    BEGIN_SINGLE_CORE
+        ${data_in_type.typeName} ref_${data_out}_${data_in} = ${data_in};
+        ${data_out_type.typeName} ref_${data_out}_${data_out} = ${data_out};
+        for (uint32_t n=0; n<${batch}; ++n) {
+            Conv1d_fp${data_in_type.referencedType.typeWidth}_fp${weight_type.referencedType.typeWidth}_fp${data_out_type.referencedType.typeWidth}(
+                ref_${data_out}_${data_in}, ${ch_im_in}, ${dim_im_in_y},
+                ${weight}, ${ch_im_out}, ${dim_kernel_y},
+                ${stride_y},
+                ${bias},
+                ${has_bias},
+                ref_${data_out}_${data_out},
+                ${dim_im_out_y}
+            );
+
+            ref_${data_out}_${data_in} += ${batchOffsetIn};
+            ref_${data_out}_${data_out} += ${batchOffsetOut};
+        }
+    END_SINGLE_CORE
+    """)