+++ Deeploy/Targets/Generic/Templates/FloatMaxPoolTemplate.py
@@ -20,3 +20,24 @@ BEGIN_SINGLE_CORE
     }
 END_SINGLE_CORE
 """)
+
+reference1DTemplate = NodeTemplate("""
+<%
+batchOffsetIn = ch_im_in * dim_im_in_y
+batchOffsetOut = ch_im_out * dim_im_out_y
+%>
+    // 1D Float MaxPool (Name: ${nodeName}, Op: ${nodeOp})
+    BEGIN_SINGLE_CORE
+        ${data_in_type.typeName} ref_${data_out}_${data_in} = ${data_in};
+        ${data_out_type.typeName} ref_${data_out}_${data_out} = ${data_out};
+        for (uint32_t n=0; n<${batch}; ++n) {
+            MaxPool1d_fp32_fp32(
+                ref_${data_out}_${data_in}, ${ch_im_in}, ${dim_im_in_y},
+                ${dim_kernel_y}, ${stride_y},
+                ref_${data_out}_${data_out}
+            );
+            ref_${data_out}_${data_in} += ${batchOffsetIn};
+            ref_${data_out}_${data_out} += ${batchOffsetOut};
+        }
+    END_SINGLE_CORE
+""")